requirements.txt
amqp==2.6.0
appdirs==1.4.3
asgiref==3.3.1
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.2.0
celery==4.4.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==2.2.4
django-cors-headers==2.1.0
djangorestframework==3.9.2
formencode==1.3.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.5
redis==3.5.3
requests==2.22.0
retrying==1.3.3
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.14.0
sqlparse==0.4.1
stripe==2.55.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
